Adjust find nearest road func for osm features to internalize processing of all road types (groups).
The find nearest func calls for each group were previously parallelize but these are relatively quick while requiring a lot of memory. Most of the time this was still okay on 16GB machine but larger countries ran out of memory or ended up processing very slowly as memory was spilled to disk.